pyproject.toml
[tool.poetry]
name = "django task manager"
version = "0.1.0"
description = "An Employee Task Manager System Built with Django"
authors = ["Ted Ngeene <tngeene27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "4.2.2"
djangorestframework = "3.14.0"
djoser = "2.1.0"
django-cors-headers = "3.13.0"
pre-commit = "3.2.0"
django-filter = "23.1"
gunicorn = "20.1.0"
psycopg2 = "2.9.6"
pillow = "9.3.0"
django-environ = "0.9.0"
django-extensions = "3.2.1"


[tool.poetry.group.test]

[tool.poetry.group.test.dependencies]
factory-boy = "3.2.1"
freezegun = "1.2.2"
pytest = "7.2.2"
pytest-django = "4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
